Changed FetchStatus to enum. Refactoring slices. -changed the data type fetchStatus in slices to enum -renamed the directory const on enum
