--- src/const/fetchStatus.ts
@@ -1,5 +1,0 @@
-export const FetchStatus = {
-  SUCCESS: 'RESPONCE',
-  FAILED: 'ERROR',
-  PENDING: 'PENDING',
-};
