+++ src/enums/fetchStatus.ts
@@ -1,0 +1,6 @@
+export enum FetchStatus {
+  DEFAULT = '',
+  SUCCESS = 'RESPONSE',
+  FAILED = 'ERROR',
+  PENDING = 'PENDING',
+}
