+++ src/modules/AuthentificationForm/components/RestorePasswordConfirm/index.tsx
@@ -3,7 +3,7 @@ import { NavLink, useRouteMatch } from "react-router-dom";
 import { BackToSignInButton } from "components";
 import { useRedirect } from "hooks";
 import { RestoreMessage } from "../RestoreMessage";
-import { AppPaths, AuthPath } from "const";
+import { AppPaths, AuthPath } from "enums";
 import { FeedbackForm } from "../../styles";
 
 export function RestorePassConfirm () {
